+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # OSGL HTTP Change Log
 
+1.13.0
+* Improve content negotiation logic #35
+
+1.13.0 
+* Improve content negotiation logic #35
+
 1.12.0 - 23/Nov/2019
 * H.Format enhancement #34
     - H.Format - use String.intern() to set name and contentType